cook/pack: harden untrusted size/count fields against overflow + over-alloc
Continues the deserialization-safety sweep into the two parsers every
cooked asset passes through before the codec layer.

cook::CookedAsset::deserialize: `kCookHeaderBytes + size` was u32
(kCookHeaderBytes is constexpr u32 16). `size` is read straight from
the blob, so a value near UINT32_MAX wrapped past the bound check and
payload.assign walked far OOB. Widened to usize (64-bit) — the outer
container guard is now overflow-safe, like the asset/codec decoders.

pack Archive::open: `entries_.reserve(entry_count)` used the raw
untrusted u32 header field — a corrupt header (entry_count = 0xFFFFFFFF)
requested a ~240 GB reserve and OOM-terminated the process. Each
on-disk entry is >= 30 bytes, so cap the pre-allocation at
(sz - index_offset) / 30; the index loop already stops at EOF, so the
real entry set is unchanged for valid archives.

Tests: cook gains a UINT32_MAX size-field overflow guard; pack patches
a valid archive's header entry_count to 0xFFFFFFFF and asserts open()
returns cleanly with its real 4 entries (process-terminating pre-fix).
cook 49 checks, pack 59 checks, all green.

Release build clean /W4 /WX, 56/56 regression green.

+++ runtime/cook/src/cook.cpp
@@ -85,7 +85,10 @@ bool CookedAsset::deserialize(const cardinal::vector<u8>& bytes, CookedAsset& ou
     out.type           = static_cast<AssetType>(rd_u32(4));
     const u32 size     = rd_u32(8);
     out.cooker_version = rd_u32(12);
-    if (kCookHeaderBytes + size > bytes.size()) return false;
+    // `size` is attacker-controlled; `kCookHeaderBytes + size` in u32
+    // wraps for size near UINT32_MAX, slips past this guard, and the
+    // payload.assign below walks far out of bounds. 64-bit arithmetic.
+    if (static_cast<usize>(kCookHeaderBytes) + size > bytes.size()) return false;
     out.payload.assign(bytes.begin() + kCookHeaderBytes,
                        bytes.begin() + kCookHeaderBytes + size);
     return true;
